Routes check
